requirements.txt
dj-database-url
django-ninja
django
python-dotenv